move scoreRepo to inmem package
