+++ command/score.go
@@ -44,45 +44,3 @@ func (c *scoreCmd) calcScore() int {
 		return c.scoreRepo.Decr(c.user)
 	}
 }
-
-type scoreRepository struct {
-	scores   map[string]int
-	lastUser string
-}
-
-type ScoreRepository interface {
-	LastUser() string
-	Incr(user string) int
-	Decr(user string) int
-	UserScore(user string) int
-}
-
-func NewScoreRepository(scores map[string]int) ScoreRepository {
-	return &scoreRepository{
-		scores: scores,
-	}
-}
-
-func (r *scoreRepository) LastUser() string {
-	return r.lastUser
-}
-
-func (r *scoreRepository) Incr(user string) int {
-	score := r.scores[user]
-	score++
-	r.scores[user] = score
-	r.lastUser = user
-	return score
-}
-
-func (r *scoreRepository) Decr(user string) int {
-	score := r.scores[user]
-	score--
-	r.scores[user] = score
-	r.lastUser = user
-	return score
-}
-
-func (r *scoreRepository) UserScore(user string) int {
-	return r.scores[user]
-}
