+++ command/score_test.go
@@ -3,27 +3,27 @@ package command
 import (
 	"testing"
 
-	"github.com/stretchr/testify/assert"
+	"github.com/golang/mock/gomock"
+
+	mock_command "github.com/johnmanjiro13/dokkoi/command/mock_score"
 )
 
 func TestScoreCmd_CalcScore(t *testing.T) {
-	scores := map[string]int{
-		"johnman": 1,
-		"625":     2,
-	}
-	repo := NewScoreRepository(scores)
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+	repo := mock_command.NewMockScoreRepository(ctrl)
 
 	tests := map[string]struct {
 		user     string
-		expected int
+		operator string
 	}{
-		"user already exists": {
+		"incr": {
 			user:     "johnman",
-			expected: 2,
+			operator: incrOperator,
 		},
 		"user not exists": {
 			user:     "kairyu",
-			expected: 1,
+			operator: decrOperator,
 		},
 	}
 
@@ -34,106 +34,12 @@ func TestScoreCmd_CalcScore(t *testing.T) {
 				user:      tt.user,
 				operator:  incrOperator,
 			}
-			actual := cmd.calcScore()
-			assert.Equal(t, tt.expected, actual)
-		})
-	}
-}
-
-func TestScoreRepository_Incr(t *testing.T) {
-	scores := map[string]int{
-		"johnman": 1,
-		"625":     2,
-	}
-	repo := NewScoreRepository(scores)
-
-	tests := map[string]struct {
-		user     string
-		expected int
-	}{
-		"user already exists": {
-			user:     "johnman",
-			expected: 2,
-		},
-		"user not exists": {
-			user:     "kairyu",
-			expected: 1,
-		},
-	}
-
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			actual := repo.Incr(tt.user)
-			assert.Equal(t, tt.expected, actual)
-			assert.Equal(t, tt.user, repo.LastUser())
-		})
-	}
-}
-
-func TestScoreRepository_Decr(t *testing.T) {
-	scores := map[string]int{
-		"johnman": 1,
-		"625":     2,
-	}
-	repo := NewScoreRepository(scores)
-
-	tests := map[string]struct {
-		user     string
-		expected int
-	}{
-		"user already exists": {
-			user:     "johnman",
-			expected: 0,
-		},
-		"user not exists": {
-			user:     "kairyu",
-			expected: -1,
-		},
-	}
-
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			actual := repo.Decr(tt.user)
-			assert.Equal(t, tt.expected, actual)
-			assert.Equal(t, tt.user, repo.LastUser())
-		})
-	}
-}
-
-func TestScoreRepository_LastUser(t *testing.T) {
-	scores := map[string]int{}
-	repo := NewScoreRepository(scores)
-	repo.Incr("johnman")
-	assert.Equal(t, "johnman", repo.LastUser())
-	repo.Decr("god")
-	assert.Equal(t, "god", repo.LastUser())
-}
-
-func TestScoreRepository_UserScore(t *testing.T) {
-	scores := map[string]int{
-		"johnman": 1,
-		"625":     2,
-	}
-	repo := NewScoreRepository(scores)
-
-	tests := map[string]struct {
-		user     string
-		expected int
-	}{
-		"user already exists": {
-			user:     "johnman",
-			expected: 1,
-		},
-		"user not exists": {
-			user:     "kairyu",
-			expected: 0,
-		},
-	}
-
-	for name, tt := range tests {
-		t.Run(name, func(t *testing.T) {
-			actual := repo.UserScore(tt.user)
-			assert.Equal(t, tt.expected, actual)
+			if cmd.operator == incrOperator {
+				repo.EXPECT().Incr(cmd.user)
+			} else if cmd.operator == decrOperator {
+				repo.EXPECT().Decr(cmd.user)
+			}
+			cmd.calcScore()
 		})
 	}
 }
