+++ command/service.go
@@ -17,6 +17,12 @@ var ErrImageNotFound = errors.New("image was not found.")
 type CustomSearchRepository interface {
 	SearchImage(query string) (*customsearch.Result, error)
 }
+type ScoreRepository interface {
+	LastUser() string
+	Incr(user string) int
+	Decr(user string) int
+	UserScore(user string) int
+}
 
 type Service interface {
 	GetCommand(content string) DokkoiCmd
